% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{benchmarkWrapper}
\alias{benchmarkWrapper}
\title{A wrapper that performs the whole benchmark analysis}
\usage{
benchmarkWrapper(rpath, ANALYSIS_NAME, qt)
}
\arguments{
\item{rpath}{The path were the quantification files are stored, and where the output files will be saved.}

\item{ANALYSIS_NAME}{The name of the analysis pipeline}

\item{qt}{A string indicating the unit of the expression matrix (either "FPKM", "TPM" or "COUNTS").}
}
\value{
Nothing but saves a bunch of files in 'rpath'.
}
\description{
Performs the whole series of benchmark analysis (see \code{\link{txLevel}}, \code{\link{compareWithNanostring}}, \code{\link{analyzeSpikein}}, \code{\link{compareWithPCR}}, and \code{\link{compareSimulated}}).
The function expects quantification files with the right column headers ("AJ80" and so on, or "s1","s2" and so on for the simulated data) to be in the rpath folder, and to bear some kind of recognizable name.
To avoid confusion, we suggest using the filename 'transcripts.quant' for transcript-level quantification, 'genes.quant' for gene-level quantification (optional), and 'simulated.quant' for transcript-level quantification of the simulated dataset (optional).
If you want to specify files manually, use the individual underlying functions.
}
\examples{
# first we create a directory and put the example quantification file in it:
data(exampledata)
dir.create("example")
write.table(exampleTranscriptLevel,"w12.transcripts.quant",col.names=T,row.names=T,sep="\\t",quote=F)
write.table(exampleGeneLevel,"w12.genes.quant",col.names=T,row.names=T,sep="\\t",quote=F)
# run the wrapper, specifying that folder:
benchmarkWrapper("example", "tophat.featureCount", qt="COUNTS")

}

