% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.simulated.R
\name{compareSimulated}
\alias{compareSimulated}
\title{Transcript-level benchmark of the simulated data}
\usage{
compareSimulated(ANALYSIS_NAME, txfile = "simulated.quant")
}
\arguments{
\item{ANALYSIS_NAME}{A string indicating the name of the analysis/pipeline. Will be used in filenames, plot titles, etc.}

\item{txfile}{The path to the transcript-level RNAseq expression matrix. If not given, will look for relevant files in the working directory. The expression matrix should have gene symbols in the first column/row.names, and sample names (either 's1', 's2', etc., or '1' '2', etc..) as column headers.}

\item{qt}{A string indicating the unit of the expression matrix (either "FPKM", "TPM" or "COUNTS").}

\item{normMethod}{The normalization method to use (see \code{\link{donorm}}). Defaults to 'TMM'.}

\item{uniquelyMappableLengths}{Logical, whether to use uniquely mappable length, rather than full length, for FPKM calculation. Default TRUE.}

\item{requireAll}{logical, whether all samples are required to proceed.}
}
\value{
Nothing, but produces many files in the working directory...
}
\description{
Loads and processes the RNAseq data before comparing it with the real foldchanges. 
Will produce a number of benchmarking plots and files in the current working directory.
}

