% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.diff.R
\name{compareSpikeinDEcalls}
\alias{compareSpikeinDEcalls}
\title{compareSpikeinDEcalls}
\usage{
compareSpikeinDEcalls(tests, thres = 0.01, colors = NULL)
}
\arguments{
\item{tests}{A list of tests, each element of which should be a data.frame with a "p" column indicating the p-value under that test, and have ERCC IDs as row.names.}

\item{thres}{Numeric value between 0 and 1, indicating the p-value treshold to use. Default 0.01.}

\item{colors}{A vector of colors to be used (R arbitrary colors if NULL), or "greys" to use different tones of grey.}
}
\value{
A layout of two plots, with ROC curves on the left and a barplot of accuracy measurements on the right.
}
\description{
Benchmarks a list of differential expression calls between the two spike-in mixes.
}

