% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.compareWithNanostring.R
\name{compareWithNanostring}
\alias{compareWithNanostring}
\title{Comparison with Nanostring}
\usage{
compareWithNanostring(ANALYSIS_NAME, rnaseq = NULL, qt, normMethod = NULL)
}
\arguments{
\item{ANALYSIS_NAME}{A string indicating the name of the analysis/pipeline. Will be used in filenames, plot titles, etc.}

\item{rnaseq}{The path to the gene-level RNAseq expression matrix. If not given, will look for relevant files in the working directory. The expression matrix should have refseq id or gene symbols in the first column/row.names, and sample names (e.g. 'AJ80') as column headers.}

\item{qt}{A string indicating the unit of the expression matrix (either "FPKM", "TPM" or "COUNTS").}

\item{normMethod}{The normalization method to use (see \code{\link{donorm}}). Defaults to 'housekeeping' for gene-level, and 'TMM' for transcript-level.}
}
\value{
Nothing, but produces many files in the working directory...
}
\description{
Loads and processes the RNAseq data before comparing it with the Nanostring quantification. 
Will produce a number of benchmarking plots and files in the current working directory.
}
\examples{
# first we create a directory and put the example quantification file in it:
data(exampledata)
dir.create("example")
write.table(exampleGeneLevel,"w12.genes.quant",sep="\\t",quote=F)
# then we run the function, giving a name to the analysis, 
# specifying the file and type of quantification:
compareWithNanostring("tophat.featureCount", "w12.genes.quant", qt="COUNTS")

}

