% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{convertTx2Genes}
\alias{convertTx2Genes}
\title{Summarizes transcript quantifications to the gene level}
\usage{
convertTx2Genes(tx)
}
\arguments{
\item{tx}{A data.frame of transcript values, with refseq IDs as row.names.}
}
\value{
A data.frame of gene values.
}
\description{
Summarizes transcript quantifications to the gene level by summing transcript values belonging to the same gene.
}

