% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.diff.R
\name{deNanostring.compare}
\alias{deNanostring.compare}
\title{deNanostring.compare}
\usage{
deNanostring.compare(results, method, norm, quantification = "",
  threshold = 0.01)
}
\arguments{
\item{results}{A dataframe as produced by the \code{\link{deNanostring}} function, with gene symbols as row names, and containing the columns "p" (p-value) and "log2FC".}

\item{method}{The differential expression method used (for reporting only)}

\item{norm}{The normalization method used (for reporting only)}

\item{quantification}{The quantification used (for reporting only)}

\item{threshold}{The p-value threshold applied on the Nanostring data to identify whether a gene is differentially-expressed or not.}
}
\value{
A layout with 3 plots
}
\description{
Compares the p-values and estimated foldchanges from a DEA analysis to a log-t-test performed on the nanostring data.
}

