% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.diff.R
\name{deNanostring}
\alias{deNanostring}
\title{deNanostring}
\usage{
deNanostring(rnaseq = NULL, method = "edgeR", norm = "TMM",
  quantification = "", threshold = 0.01)
}
\arguments{
\item{rnaseq}{Either the (gene-level) count matrix, or a character indicating the location of the file.}

\item{method}{The differential expression method to use. Either 'edgeR', 'DESeq', 'DESeq2', 'EBSeq', 'voom', 't' (t-test), or 'logt' (t-test on log-transformed values).}

\item{norm}{The normalization method to use (either "linear", or any of the methods supported by edgeR's \code{\link[edgeR]{calcNormFactors}}. Defaults to 'TMM'.}

\item{quantification}{A string indicating the quantification that was used to produce the expression matrix (use for labeling)}

\item{threshold}{The p-value threshold applied on the Nanostring data to identify whether a gene is differentially-expressed or not.}
}
\value{
A deNanostring.compare plot
}
\description{
Runs a differential expression analysis and compares it to a log-t-test performed on the nanostring data.
}
\examples{
data(exampledata)
deNanostring(exampleGeneLevel, method="edgeR", norm="TMM", 
  quantification="Tophat-featureCounts")

}

