% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.diff.R
\name{deSpikein.compare}
\alias{deSpikein.compare}
\title{deSpikein.compare}
\usage{
deSpikein.compare(d, quantification, norm = "", method = "")
}
\arguments{
\item{d}{A data.frame with at least the columns "p" (for p-value) and "log2FC", as produced in the \code{\link{differentialExpression}} function.}

\item{quantification}{Name of the pipeline that produced the underlying quantification, used in plot titles.}

\item{norm}{Normalization method, used in plot titles.}

\item{method}{DEA method, used in plot titles.}
}
\value{
Nothing.
}
\description{
A wrapper to create spike-in DEA benchmarking plots
}

