% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.diff.R
\name{deSpikein}
\alias{deSpikein}
\title{Runs a spike-in differential expression analysis.}
\usage{
deSpikein(dat, method = "edgeR", norm = "TMM", quantification = "",
  homogenize.mixes = T, saveResults = F, savePlot = F)
}
\arguments{
\item{dat}{The expression matrix or data.frame, with gene symbols or transcript Refseq IDs as row.names, and sample names as column headers.}

\item{method}{The differential expression method to use. Either 'edgeR', 'DESeq', 'DESeq2', 'EBSeq', 'voom', 't' (t-test), or 'logt' (t-test on log-transformed values). A string indicating the unit of the expression matrix (either "FPKM", "TPM" or "COUNTS").}

\item{norm}{The normalization method to use (see \code{\link{donorm}}). Defaults to 'TMM'.}

\item{quantification}{A string indicating the name of the analysis/pipeline form which the quantification comes. Will be used in filenames, plot titles, etc.}

\item{homogenize.mixes}{logical, whether the two spike-in mixes should be homogenized for the purpose of calculating normalization factors.}

\item{saveResults}{Logical, whether to save the results of the DEA in the current working directory (default FALSE).}

\item{savePlot}{Logical, whether to save the plot in the current working directory (default FALSE).}
}
\value{
A data.frame with the results of the differential expression analysis, as well as a plot.
}
\description{
Runs a differential expression analysis and compares it to real differences between spike-in mixes. 
Sleuth is handled in a different function (see \code{\link{sleuthWrapper}})
}
\examples{
data(exampledata)
res <- deSpikein(exampleGeneLevel, method="edgeR", norm="TMM", quantification="Tophat-featureCounts")

}

