% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{donorm}
\alias{donorm}
\title{normalizes a dataset}
\usage{
donorm(dataset, method = "TMM")
}
\arguments{
\item{dataset}{is a data.frame}

\item{method}{is a character string of either 'linear', 'housekeeping', 'quantile', or any method supported by edgeR's \code{\link[edgeR]{calcNormFactors}}.}
}
\value{
The normalized data.frame.
}
\description{
Wrapper to normalize a dataset, using the given method.
}
\examples{
donorm(matrix(1:12,nrow=4),"linear")

}

