% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{fc2mean}
\alias{fc2mean}
\title{foldchange-to-the-mean}
\usage{
fc2mean(x)
}
\arguments{
\item{x}{a numeric matrix or data.frame.}
}
\value{
The matrix of corresponding foldchanges to the row's mean.
}
\description{
Returns the foldchange to the row's mean for each value.
}
\examples{
fc2mean(matrix(1:12,nrow=3))

}

