% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.diff.R
\name{fcdev}
\alias{fcdev}
\title{Plots deviation from real foldchange by p-value}
\usage{
fcdev(d, real.log2fc, title = "")
}
\arguments{
\item{d}{A data.frame as produced in the \code{\link{differentialExpression}} function.}

\item{real.log2fc}{A vector of length nrow(d) containing the real foldchanges.}

\item{title}{Plot title.}
}
\value{
Nothing.
}
\description{
Plots deviation from real foldchange by p-value
}

