% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{foldchange}
\alias{foldchange}
\title{foldchange}
\usage{
foldchange(x, y, na.fc = NA)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of same length as x.}

\item{na.fc}{the value to return is both x and y are 0}
}
\value{
A vector of same length as x containing the foldchange of y/x
}
\description{
Returns the foldchange, or the non-0 value +1 (or its inverse) if one of the values is 0, or 1/NA (as specified) if both are zero.
}
\examples{
foldchange( c(5,Inf,0,6,0), c(0,1,2,3,0) )

}

