% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getLinearNormalizers}
\alias{getLinearNormalizers}
\title{getLinearNormalizers}
\usage{
getLinearNormalizers(dataset, tryrobust = F)
}
\arguments{
\item{dataset}{a numeric data.frame or matrix.}

\item{tryrobust}{logical, whether to try to fit a robust linear model instead of a normal one.}
}
\value{
A vector of length ncol(dataset) containing the scale factors of each column of dataset.
}
\description{
Returns linear scale factors for each column of the dataset, using the first column as a reference and fitting a linear model between each other colum and the first.
}
\examples{
getLinearNormalizers(matrix(1:12,nrow=3))

}

