% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getLogFC}
\alias{getLogFC}
\title{Retrieves the log2(foldchange)}
\usage{
getLogFC(x, groups)
}
\arguments{
\item{x}{a numeric vector.}

\item{groups}{a vector of same length as x, indicating to which of two groups each value belongs. If more than two groups, only the first two are used.}
}
\value{
Returns the log2 foldchange, or the log2 of the largest mean if one of the mean is 0, or 0 if both means are 0.
}
\description{
Retrieves the log2(foldchange).
}

