% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.diff.R
\name{homomixes}
\alias{homomixes}
\title{Homogenizes the two spike-in mixes.}
\usage{
homomixes(sp, mix1 = NULL)
}
\arguments{
\item{sp}{Expression matrix or data.frame, with spike-in IDs as row.names.}

\item{mix1}{The column headers of the samples that were spiked with mix1 (optional). This is detected automatically for the datasets included in this study.}
}
\value{
A data.frame including only the spike-ins, and with the two mixes 'homogenized'.
}
\description{
Divides spike-ins quantifications by their expected foldchange, thereby making all samples in principle equal. This is used for the purpose of calculating normalization factors.
}

