% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{isCleanData}
\alias{isCleanData}
\title{isCleanData}
\usage{
isCleanData(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector of same length as x}
}
\value{
indices that are 'clean' (non-missing, non-infinite and non-NaN) in both x and y.
}
\description{
Returns which couples of x and y values are clean (non-missing, non-infinite and non-NaN)
}
\examples{
isCleanData(c(2,NA,3),c(0,5,NA))

}

