% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{MdAE}
\alias{MdAE}
\title{Median absolute error}
\usage{
MdAE(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector of same length as x.}
}
\value{
The median absolute error between x and y.
}
\description{
Returns the median absolute error between vectors x and y.
}
\examples{
MdAE(1:3,jitter(1:3))

}

