% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.diff.R
\name{mROC}
\alias{mROC}
\title{mROC}
\usage{
mROC(tests, sig, show.AUC = T, thres = 0.01, lwd = 2, colors = NULL,
  ...)
}
\arguments{
\item{tests}{A list of tests, each element of which should be a data.frame with a "p" column indicating the p-value under that test.}

\item{sig}{A logical vector indicating whether each row of elements in 'tests' is actually differentially-expressed (TRUE) or not. If this is a named vector, the names will be used to order the rows in 'tests', otherwise they are assumed to be ordered in the same way.}

\item{show.AUC}{Logical; whether to show the area under the curve in the figure's legend. Default TRUE.}

\item{thres}{Numeric value between 0 and 1, indicating the p-value treshold to be plotted on each curve. Set to NULL to plot no threshold.}

\item{lwd}{Line width, passed to the plotting functions. Default 2.}

\item{...}{Any further argument passed to the initial plot function.}
}
\description{
Plots superposed ROC curves.
}

