% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.diff.R
\name{myc}
\alias{myc}
\title{Plots sensitivity and specificity by p-value}
\usage{
myc(logp, sig, title = "", xlab = "log10(p-value)")
}
\arguments{
\item{logp}{The log10(p-value) of each feature.}

\item{sig}{A logical vector indicating whether each element in logp is actually differentially-expressed (TRUE) or not.}

\item{xlab}{Plot xlab, defaults to "log10(p-value)".}
}
\value{
Nothing.
}
\description{
Plots sensitivity and specificity by p-value
}

