% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plGetModel}
\alias{plGetModel}
\title{Retrieves a linear model, if possible.}
\usage{
plGetModel(x, y, no.intercept = T, tryrobust = F)
}
\arguments{
\item{x}{a numeric vector containing the independent variable.}

\item{y}{a numeric vector of same length as x, containing the dependent variable.}

\item{no.intercept}{logical, indicating whether to disable the intersect in the linear model (defaults TRUE, i.e. models must go through the origin)}

\item{tryrobust}{logical, indicating whether to try to fit a robust linear model (see \code{\link[MASS]{rlm}}) instead of a normal one.}
}
\value{
Returns the linear model, if the fit was possible, NA otherwise.
}
\description{
Retrieves a linear model, if possible.
}

