% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.diff.R
\name{pval2fc}
\alias{pval2fc}
\title{Plots p-values according to real foldchange.}
\usage{
pval2fc(d, real.log2fc, jitter.arrows = TRUE, title = "")
}
\arguments{
\item{d}{A data.frame as produced in the \code{\link{differentialExpression}} function.}

\item{real.log2fc}{A vector of length nrow(d) containing the real foldchanges.}

\item{jitter.arrows}{Logical, whether the position of arrows outside the plotting range should be jittered to better see their number.}

\item{title}{Plot title.}
}
\value{
Nothing.
}
\description{
Plots p-values according to real foldchange.
}

