% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{readBenchmarkResults}
\alias{readBenchmarkResults}
\title{readBenchmarkResults}
\usage{
readBenchmarkResults(folders, dataset = "w12")
}
\arguments{
\item{folders}{A character vector containing the paths to the benchmark folders}

\item{dataset}{Compare metrics for which dataset - either "w12" (default) or "w6"}
}
\value{
A dataframe, with metrics as columns and analyses (i.e. folders) as rows. For more information about the metrics, see the Suppl. Table 2 of the paper.
}
\description{
Reads results metrics of several benchmark folders (i.e. output folders of \code{\link{benchmarkWrapper}}) into a matrix, for comparison.
}

