\name{RNAontheBENCH-package}
\alias{RNAontheBENCH-package}
\alias{RNAontheBENCH}
\docType{package}
\title{
\packageTitle{RNAontheBENCH}
}
\description{
RNAontheBENCH - Benchmark of RNAseq quantification and DEA
}
\details{
RNA on the Benchmark of Expression by nCounter Hybridization (RNAontheBENCH) is a resource for benchmarking RNAseq quantification and differential expression analysis methods on the basis of empirical data. For more information, see \code{vignette("RNAontheBENCH")}.
}
\author{
\packageAuthor{RNAontheBENCH}

Maintainer: \packageMaintainer{RNAontheBENCH}
}