% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.diff.R
\name{ROC}
\alias{ROC}
\title{ROC}
\usage{
ROC(p, sig, main = NULL)
}
\arguments{
\item{p}{The p-value.}

\item{sig}{A logical vector indicating whether each element in 'p' is actually differentially-expressed (TRUE) or not.}

\item{main}{Plot title.}
}
\value{
Nothing.
}
\description{
Plots the ROC curve.
}

