% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.diff.R
\name{runTests}
\alias{runTests}
\title{Runs a differential expression analysis.}
\usage{
runTests(data, groups, norm = "TMM", de = "edgeR", homogenize.mixes = T)
}
\arguments{
\item{data}{The expression matrix or data.frame, with gene symbols or transcript Refseq IDs as row.names.}

\item{groups}{A logical vector (or coercible to logical) of length ncol(data), indicating to which group each sample (i.e. column in 'data') belongs. There can be only two groups.}

\item{norm}{The normalization method to use (either "linear", or any of the methods supported by edgeR's \code{\link[edgeR]{calcNormFactors}}. Defaults to 'TMM'.}

\item{de}{The differential expression method to use. Either 'edgeR', 'DESeq', 'DESeq2', 'EBSeq', 'voom', 't' (t-test), or 'logt' (t-test on log-transformed values). A string indicating the unit of the expression matrix (either "FPKM", "TPM" or "COUNTS").}

\item{homogenize.mixes}{logical, whether the two spike-in mixes should be homogenized for the purpose of calculating normalization factors.}
}
\value{
A data.frame with the results of the differential expression analysis.
}
\description{
Runs a differential expression analysis using the selected method. This is used by the \code{\link{differentialExpression}} function. 
This is simply a wrapper to make sure that the different analysis methods use the same normalization method and return results in the same format.
}

