% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqc.R
\name{seqc.diff.example}
\alias{seqc.diff.example}
\title{seqc.diff.example}
\usage{
seqc.diff.example(e = NULL, site = "BGI", tests = c("edgeR", "voom",
  "DESeq2"), between.groups = 1:5, inner.groups = list(c(1, 2, 3), c(4, 5)),
  do.plot = T, returnData = F)
}
\arguments{
\item{e}{The count matrix or data.frame. Anything can be used as row names (e.g. gene symbols, transcript ids...). If NULL, the 'site' parameter is used to fetch the quantification from the seqc package (if installed).}

\item{site}{If 'e' is NULL, data from the specified sequencing site will be fetched from the seqc package. Default BGI.}

\item{tests}{A vector of the DEA methods to be compared.}

\item{between.groups}{A vector of integers from 1 to 5 indicating the replicates to be used for comparison across conditions. Default 1:5.}

\item{inner.groups}{A list of two vectors, each containing the replicates to be used for comparison inside each condition. Default list(c(1,2,3),c(4,5)).}

\item{do.plot}{Logical, whether to produce Positives plots (see posplot). Default TRUE.}

\item{returnData}{Logical, whether to return the resulting data.}
}
\value{
Produces a plot, and returns a list if returnData=T.
}
\description{
Compares differential expression analysis (DEA) methods using the 'diff.seqc' function, and produces ROC-like plots.
}

