% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqc.R
\name{seqc.diff}
\alias{seqc.diff}
\title{seqc.diff}
\usage{
seqc.diff(e = NULL, norm = "TMM", de = "edgeR", site = "BGI",
  between.groups = 1:5, inner.groups = list(c(1, 2, 3), c(4, 5)),
  do.plot = TRUE)
}
\arguments{
\item{e}{The count matrix or data.frame. Anything can be used as row names (e.g. gene symbols, transcript ids...). If NULL, the 'site' parameter is used to fetch the quantification from the seqc package (if installed).}

\item{norm}{The normalization method to use (see \code{\link{donorm}}). Defaults to 'TMM'.}

\item{de}{The differential expression method to use. Either 'edgeR', 'DESeq', 'DESeq2', 'EBSeq', 'voom', 't' (t-test), or 'logt' (t-test on log-transformed values).}

\item{site}{If 'e' is NULL, data from the specified sequencing site will be fetched from the seqc package. Default BGI.}

\item{between.groups}{A vector of integers from 1 to 5 indicating the replicates to be used for comparison across conditions. Default 1:5.}

\item{inner.groups}{A list of two vectors, each containing the replicates to be used for comparison inside each condition. Default list(c(1,2,3),c(4,5)).}

\item{do.plot}{Logical, whether to produce Positives plots (see posplot). Default TRUE.}
}
\value{
A list of differential expression calls.
}
\description{
Runs a comparison of differential expression calls on the SEQC data, comparing calls across conditions to calls within conditions. 
The assumption here (and made by the SEQC consortium) is that differential expression across replicates are false positives.
See \code{\link{seqc.diff.example}} for an example usage.
}

