% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqc.R
\name{seqc.diff.sleuth}
\alias{seqc.diff.sleuth}
\title{seqc.diff.sleuth}
\usage{
seqc.diff.sleuth(folders, between.groups = 1:5, inner.groups = list(c(1, 2,
  3), c(4, 5)), do.plot = T)
}
\arguments{
\item{folders}{Folders containing the quantification and bootstraps. The folders should be named "A_1" and so on.}

\item{between.groups}{A vector of integers from 1 to 5 indicating the replicates to be used for comparison across conditions. Default 1:5.}

\item{inner.groups}{A list of two vectors, each containing the replicates to be used for comparison inside each condition. Default list(c(1,2,3),c(4,5)).}

\item{do.plot}{Logical, whether to produce Positives plots (see posplot). Default TRUE.}
}
\value{
A list of differential expression calls.
}
\description{
Sleuth wrapper for the analysis of SEQC data.
}

