% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqc.R
\name{seqc.get.stats}
\alias{seqc.get.stats}
\title{seqc.get.stats}
\usage{
seqc.get.stats(ps, use.fdr = FALSE, pthreshold = 0.01)
}
\arguments{
\item{ps}{A list with softwares as names, and lists as elements, as produced by the 'diff.seqc.example' function. Each sublist contains the results of comparisons between  A vs B, C vs D, and within each group.}

\item{use.fdr}{Logical; whether to use adjust the p-values for multiple testing (FDR); default FALSE.}

\item{pthreshold}{P-value (or FDR if use.fdr=TRUE) threshold to use to compute test statistics.}
}
\value{
A data frame.
}
\description{
Returns test statistics for a list of differential expression calls on the SEQC data.
}

