% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqc.R
\name{seqc.prepareData}
\alias{seqc.prepareData}
\title{seqc.prepareData}
\usage{
seqc.prepareData(site = "BGI", removeERCC = T)
}
\arguments{
\item{site}{Sequencing site, default BGI.}

\item{removeERCC}{Logical; whether to remove the spike-ins (default TRUE)}
}
\value{
A data.frame of counts, with gene symbols as row.names.
}
\description{
Fetches and prepare SEQC data using the \code{\link{seqc}} package.
}

