% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth_sailfish.R
\name{sf_to_hdf5}
\alias{sf_to_hdf5}
\title{Convert sailfish results for one sample to kallisto HDF5
TAKEN FROM SAILFISH/SALMON DEVELOPERS COMMIT ON THE SLEUTH GIT}
\usage{
sf_to_hdf5(sf_dir)
}
\arguments{
\item{sf_dir}{path to a sailfish output directory}
}
\description{
Convert sailfish results for one sample to kallisto HDF5
TAKEN FROM SAILFISH/SALMON DEVELOPERS COMMIT ON THE SLEUTH GIT
}

