% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{simStats}
\alias{simStats}
\title{Writes a table of metrics comparing two vectors}
\usage{
simStats(x, y, ANALYSIS_NAME, fname, clean = T, norm = F, incZeros = F)
}
\arguments{
\item{x}{A numeric vector of observed values.}

\item{y}{A numeric vector of expected values (same length as 'x').}

\item{ANALYSIS_NAME}{The name of the analysis pipeline}

\item{fname}{The filename for saving the metrics}

\item{clean}{Logical, whether to clean the data (default TRUE), see \code{\link{isCleanData}}.}

\item{norm}{Logical, whether to linearly normalize the two vectors before comparison (default FALSE).}

\item{incZeros}{Logical, whether to keep 0 values (default TRUE).}
}
\value{
Nothing.
}
\description{
Writes a table of metrics comparing two vectors
}

