% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth_sailfish.R
\name{sleuth_prep_sailfish}
\alias{sleuth_prep_sailfish}
\title{Constructor for a 'sleuth' object using sailfish data}
\usage{
sleuth_prep_sailfish(sf_dirs, sample_to_covariates, full_model,
  filter_fun = basic_filter, target_mapping = NULL, max_bootstrap = NULL,
  ...)
}
\arguments{
\item{sf_dirs}{a character vector of length greater than one where each
string points to a sailfish output directory}
}
\description{
Converts sailfish results to kallisto HDF5
format, then return the results of \code{\code{\link[sleuth]{sleuth_prep}}} on
the converted data.
TAKEN FROM SAILFISH/SALMON DEVELOPERS COMMIT ON THE SLEUTH GIT
}
\details{
See \code{\code{\link[sleuth]{sleuth_prep}}} for parameters other than sf_dirs.
}

