% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.diff.R
\name{sleuthWrapper}
\alias{sleuthWrapper}
\title{Runs a spike-in differential expression analysis using sleuth.}
\usage{
sleuthWrapper(name, folders = NULL, norm = "TMM", savePlot = F)
}
\arguments{
\item{name}{A string indicating the name of the analysis/pipeline form which the quantification comes. Will be used in filenames, plot titles, etc.}

\item{norm}{The normalization method to use (see \code{\link{donorm}}). Defaults to 'TMM'.}

\item{savePlot}{Logical, whether to save the plot in the current working directory (default TRUE).}

\item{folder}{A character vector containing the path to the Salmon/Sailfish/Kallisto quantification folders. If missing, will look for names matching "AJ" in the current working directory.}
}
\value{
A data.frame with the results of the differential expression analysis, as well as a plot.
}
\description{
Runs a differential expression analysis using sleuth and compares it to real differences between spike-in mixes. Currently only for the 12-samples dataset.
}

