% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{zscore}
\alias{zscore}
\title{z-score}
\usage{
zscore(x)
}
\arguments{
\item{x}{a numeric matrix or data.frame.}
}
\value{
The matrix of corresponding row z-scores.
}
\description{
Wrapper to calculate the z-score
}
\examples{
zscore(matrix(1:12,nrow=3))

}

