% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth_sailfish.R
\name{package_sf_as_kal}
\alias{package_sf_as_kal}
\title{Convert sailfish results for one or more samples to kallisto HDF5
TAKEN FROM SAILFISH/SALMON DEVELOPERS COMMIT ON THE SLEUTH GIT}
\usage{
package_sf_as_kal(sf_dirs)
}
\arguments{
\item{sf_dirs}{a character vector of length greater than one where each
string points to a sailfish output directory}
}
\description{
Convert sailfish results for one or more samples to kallisto HDF5
TAKEN FROM SAILFISH/SALMON DEVELOPERS COMMIT ON THE SLEUTH GIT
}

