% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.diff.R
\name{ROC}
\alias{ROC}
\title{ROC}
\usage{
ROC(p, sig, rounding = 5, main = NULL)
}
\arguments{
\item{p}{The p-value.}

\item{sig}{A logical vector indicating whether each element in 'p' is actually differentially-expressed (TRUE) or not.}

\item{rounding}{The rounding (number of digits) of log-transformed p-values at which to plot (default 3). Increasing this number will increase the resolution of the curve, but also increase execution time.}

\item{main}{Plot title.}
}
\value{
Nothing.
}
\description{
Plots the ROC curve.
}

