% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth_sailfish.R
\name{sf_to_hdf5}
\alias{sf_to_hdf5}
\title{Convert sailfish results for one sample to kallisto HDF5}
\usage{
sf_to_hdf5(sf_dir)
}
\arguments{
\item{sf_dir}{path to a sailfish output directory}
}
\description{
TAKEN FROM SAILFISH/SALMON DEVELOPERS COMMIT ON THE SLEUTH GIT and provided
as is; users are advised to double-check for updates on how to use salmon/sailfish
data with sleuth.
}

